Move RequestShortener tampering to stableids
