+++ webpack/module-minifier-plugin/src/StableMinifierIdsPlugin.ts
@@ -4,6 +4,7 @@
 import { compilation, Compiler } from 'webpack';
 import { createHash } from 'crypto';
 import { Tap } from 'tapable';
+import * as RequestShortener from 'webpack/lib/RequestShortener';
 
 import { STAGE_AFTER, STAGE_BEFORE } from './Constants';
 import { INormalModuleFactoryModuleData, IExtendedModule } from './ModuleMinifierPlugin.types';
@@ -36,6 +37,26 @@ export class StableMinifierIdsPlugin {
   }
 
   public apply(compiler: Compiler): void {
+    // Ensure that "EXTERNAL MODULE: " comments are portable and module version invariant
+    const baseShorten: (request: string) => string = RequestShortener.prototype.shorten;
+    RequestShortener.prototype.shorten = function (this: RequestShortener, request: string): string {
+      const baseResult: string = baseShorten.call(this, request);
+      const nodeModules: '/node_modules/' = '/node_modules/';
+
+      if (!baseResult) {
+        return baseResult;
+      }
+
+      const nodeModulesIndex: number = baseResult.lastIndexOf(nodeModules);
+      if (nodeModulesIndex < 0) {
+        return baseResult;
+      }
+
+      const nodeModulePath: string = baseResult.slice(nodeModulesIndex + nodeModules.length);
+      this.cache.set(request, nodeModulePath);
+      return nodeModulePath;
+    }
+
     const nameByResource: Map<string | undefined, string> = new Map();
     const {
       _stableIdToFinalId: stableIdToFinalId
